feat: update CI workflow for kubex_go_release with enhanced build and upload steps
