+++ docs/assets/cli_print.png
