+++ support/scripts/install.sh
