+++ version/CLI_VERSION
@@ -1,1 +1,1 @@
-v1.1.8
+v1.2.0
