+++ version/semantic.go
@@ -1,20 +1,31 @@
 package version
 
 import (
+	lgr "github.com/faelmori/logz/logger"
+
+	"github.com/spf13/cobra"
+
+	_ "embed"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"strconv"
 	"strings"
+	"time"
 )
 
-type VersionService interface {
+const moduleAlias = "Logz"
+const moduleName = "logz"
+const gitModelUrl = "https://github.com/faelmori/" + moduleName + ".git"
+const currentVersionFallback = "v1.2.0" // First version with the version file
+
+type Service interface {
 	GetLatestVersion() (string, error)
 	GetCurrentVersion() string
 	IsLatestVersion() (bool, error)
 }
-type VersionServiceImpl struct {
+type ServiceImpl struct {
 	gitModelUrl    string
 	latestVersion  string
 	currentVersion string
@@ -22,6 +33,8 @@ type VersionServiceImpl struct {
 type Tag struct {
 	Name string `json:"name"`
 }
+
+var l = lgr.NewLogger("Logz")
 
 func getLatestTag(repoURL string) (string, error) {
 	apiURL := fmt.Sprintf("%s/tags", repoURL)
@@ -49,7 +62,7 @@ func getLatestTag(repoURL string) (string, error) {
 	return tags[0].Name, nil
 }
 
-func (v *VersionServiceImpl) updateLatestVersion() error {
+func (v *ServiceImpl) updateLatestVersion() error {
 	repoURL := "https://api.github.com/repos/faelmori/spidergo"
 	tag, err := getLatestTag(repoURL)
 	if err != nil {
@@ -58,7 +71,7 @@ func (v *VersionServiceImpl) updateLatestVersion() error {
 	v.latestVersion = tag
 	return nil
 }
-func (v *VersionServiceImpl) vrsCompare(v1, v2 []int) (int, error) {
+func (v *ServiceImpl) vrsCompare(v1, v2 []int) (int, error) {
 	if len(v1) != len(v2) {
 		return 0, fmt.Errorf("version length mismatch")
 	}
@@ -75,7 +88,7 @@ func (v *VersionServiceImpl) vrsCompare(v1, v2 []int) (int, error) {
 	}
 	return 0, nil
 }
-func (v *VersionServiceImpl) versionAtMost(versionAtMostArg, max []int) (bool, error) {
+func (v *ServiceImpl) versionAtMost(versionAtMostArg, max []int) (bool, error) {
 	if comp, err := v.vrsCompare(versionAtMostArg, max); err != nil {
 		return false, err
 	} else if comp == 1 {
@@ -83,7 +96,7 @@ func (v *VersionServiceImpl) versionAtMost(versionAtMostArg, max []int) (bool, e
 	}
 	return true, nil
 }
-func (v *VersionServiceImpl) parseVersion(versionToParse string) []int {
+func (v *ServiceImpl) parseVersion(versionToParse string) []int {
 	version := make([]int, 3)
 	for idx, vStr := range strings.Split(versionToParse, ".") {
 		vS, err := strconv.Atoi(vStr)
@@ -95,7 +108,7 @@ func (v *VersionServiceImpl) parseVersion(versionToParse string) []int {
 	return version
 }
 
-func (v *VersionServiceImpl) IsLatestVersion() (bool, error) {
+func (v *ServiceImpl) IsLatestVersion() (bool, error) {
 	if v.latestVersion == "" {
 		if err := v.updateLatestVersion(); err != nil {
 			return false, err
@@ -116,7 +129,7 @@ func (v *VersionServiceImpl) IsLatestVersion() (bool, error) {
 	}
 	return false, nil
 }
-func (v *VersionServiceImpl) GetLatestVersion() (string, error) {
+func (v *ServiceImpl) GetLatestVersion() (string, error) {
 	if v.latestVersion == "" {
 		if err := v.updateLatestVersion(); err != nil {
 			return "", err
@@ -125,33 +138,106 @@ func (v *VersionServiceImpl) GetLatestVersion() (string, error) {
 
 	return v.latestVersion, nil
 }
-func (v *VersionServiceImpl) GetCurrentVersion() string { return v.currentVersion }
+func (v *ServiceImpl) GetCurrentVersion() string { return v.currentVersion }
 
-func NewVersionService() VersionService {
-	return &VersionServiceImpl{
+func NewVersionService() Service {
+	return &ServiceImpl{
 		gitModelUrl:    gitModelUrl,
 		currentVersion: currentVersion,
 		latestVersion:  "",
 	}
 }
 
-func CheckVersion() {
-	v := NewVersionService()
-	if isLatest, err := v.IsLatestVersion(); err != nil {
-		fmt.Printf("❌ Erro ao verificar versão: %v\n", err)
-	} else if isLatest {
-		fmt.Println("✅ Você está na última versão!")
+var (
+	versionCmd = &cobra.Command{
+		Use:   "version",
+		Short: "Print the version number of " + moduleAlias,
+		Long:  "Print the version number of " + moduleAlias,
+		Run: func(cmd *cobra.Command, args []string) {
+			GetVersionInfo()
+		},
+	}
+	subLatestCmd = &cobra.Command{
+		Use:   "latest",
+		Short: "Print the latest version number of " + moduleAlias,
+		Long:  "Print the latest version number of " + moduleAlias,
+		Run: func(cmd *cobra.Command, args []string) {
+			GetLatestVersionInfo()
+		},
+	}
+	subCmdCheck = &cobra.Command{
+		Use:   "check",
+		Short: "Check if the current version is the latest version of " + moduleAlias,
+		Long:  "Check if the current version is the latest version of " + moduleAlias,
+		Run: func(cmd *cobra.Command, args []string) {
+			GetVersionInfoWithLatestAndCheck()
+		},
+	}
+)
+
+//go:embed CLI_VERSION
+var currentVersion string
+
+func GetVersion() string {
+	if currentVersion == "" {
+		return currentVersionFallback
+	}
+	return currentVersion
+}
+
+func GetGitModelUrl() string {
+	return gitModelUrl
+}
+
+func GetVersionInfo() string {
+	l.Info("Version: "+GetVersion(), map[string]interface{}{})
+	l.Info("Git repository: "+GetGitModelUrl(), map[string]interface{}{})
+	return fmt.Sprintf("Version: %s\nGit repository: %s", GetVersion(), GetGitModelUrl())
+}
+
+func GetLatestVersionFromGit() string {
+	netClient := &http.Client{
+		Timeout: time.Second * 10,
+	}
+
+	gitUrlWithoutGit := strings.TrimSuffix(gitModelUrl, ".git")
+
+	response, err := netClient.Get(gitUrlWithoutGit + "/releases/latest")
+	if err != nil {
+		l.Error("Error fetching latest version: "+err.Error(), map[string]interface{}{})
+		l.Error("Url: "+gitUrlWithoutGit+"/releases/latest", map[string]interface{}{})
+		return err.Error()
+	}
+
+	if response.StatusCode != 200 {
+		l.Error("Error fetching latest version: "+response.Status, map[string]interface{}{})
+		l.Error("Url: "+gitUrlWithoutGit+"/releases/latest", map[string]interface{}{})
+		body, _ := io.ReadAll(response.Body)
+		return fmt.Sprintf("Error: %s\nResponse: %s", response.Status, string(body))
+	}
+
+	tag := strings.Split(response.Request.URL.Path, "/")
+
+	return tag[len(tag)-1]
+}
+
+func GetLatestVersionInfo() string {
+	l.Info("Latest version: "+GetLatestVersionFromGit(), map[string]interface{}{})
+	return "Latest version: " + GetLatestVersionFromGit()
+}
+
+func GetVersionInfoWithLatestAndCheck() string {
+	if GetVersion() == GetLatestVersionFromGit() {
+		l.Info("You are using the latest version.", map[string]interface{}{})
+		return fmt.Sprintf("You are using the latest version.\n%s\n%s", GetVersionInfo(), GetLatestVersionInfo())
 	} else {
-		latestV, latestVErr := v.GetLatestVersion()
-		if latestVErr != nil {
-			fmt.Printf("❌ Erro ao obter última versão: %v\n", latestVErr)
-			return
-		}
-		fmt.Printf("🔴 Nova versão disponível: %s\n", latestV)
+		l.Warn("You are using an outdated version.", map[string]interface{}{})
+		return fmt.Sprintf("You are using an outdated version.\n%s\n%s", GetVersionInfo(), GetLatestVersionInfo())
 	}
 }
-func Version() string { return NewVersionService().GetCurrentVersion() }
 
-func main() {
-
+func CliCommand() *cobra.Command {
+	versionCmd.AddCommand(subLatestCmd)
+	versionCmd.AddCommand(subCmdCheck)
+	return versionCmd
 }
